Organised the code a little
