+++ src/app/auth/registration/registration.component.ts
@@ -1,4 +1,4 @@
-import { AuthService } from './../auth.service';
+import { AuthRestService } from '../rest/auth.rest.service';
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
@@ -14,7 +14,7 @@ export class RegistrationComponent implements OnInit {
   registrationForm: FormGroup;
   errorMessage;
 
-  constructor(private authService: AuthService, private router: Router) { }
+  constructor(private authRestService: AuthRestService, private router: Router) { }
 
   ngOnInit() {
     this.registrationForm = new FormGroup({
@@ -29,7 +29,7 @@ export class RegistrationComponent implements OnInit {
 
   onSubmit() {
     console.log(this.registrationForm);
-    this.authService.createAccount(this.registrationForm.value).subscribe(
+    this.authRestService.createAccount(this.registrationForm.value).subscribe(
       (data) => {
         console.log(data);
         this.router.navigate(['/register', 'confirmation']);
