+++ src/app/rest/project.rest.ts
@@ -1,21 +1,21 @@
-import { AuthRestService } from './../shared/auth-rest.service';
+
+import { UserSessionService } from '../auth/services/user.session.service';
 import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions } from '@angular/http';
+import { Http, Headers } from '@angular/http';
 
 
 @Injectable()
 export class ProjectRest {
-    private headers: Headers;
     
-    constructor(private authRestService: AuthRestService, private http: Http) {
-        this.headers = new Headers();
-        this.headers.append('Content-Type', 'application/json');
-        let username: string = 'username';
-        let password: string = 'password';
-        this.headers.append("Authorization", "Basic " + btoa(username + ":" + password));
+    constructor(
+        private userSessionService: UserSessionService, 
+        private http: Http) {
     }
 
     getAllProjectStubs() {
-        return this.http.get('http://localhost:8080/projects/stubs/all', {headers: this.headers});
+        return this.http.get(
+            'http://localhost:8080/projects/stubs/all', 
+            {headers: this.userSessionService.getSessionAuthHeaders()}
+        );
     }
 }
