+++ src/app/services/project.service.ts
@@ -1,4 +1,3 @@
-import { AuthRestService } from './../shared/auth-rest.service';
 import { ProjectRest } from '../rest/project.rest';
 import { Injectable } from '@angular/core';
 import { Http, Headers } from '@angular/http';
@@ -6,13 +5,7 @@ import { Http, Headers } from '@angular/http';
 
 @Injectable()
 export class ProjectService {
-    private headers: Headers;
-    constructor(
-        private authRestService: AuthRestService, 
-        private projectRest: ProjectRest,
-        private http: Http) {
-        this.headers = new Headers();
-        this.headers.append('Content-Type', 'application/json');
+    constructor(private projectRest: ProjectRest) {
     }
 
     getAllProjectStubs() {
