--- src/app/shared/auth-rest.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable()
-export class AuthRestService {
-
-    constructor() {}
-
-}
